chore: refactor column visibility feature (#5638)
* Refactoring ColumnVisibility file into its own folder with separated types and utils

* prettier commit

* re-add nxCloudAccessToken

---------
