+++ packages/table-core/src/core/table.ts
@@ -29,7 +29,7 @@ import { ColumnGrouping } from '../features/ColumnGrouping'
 import { ColumnOrdering } from '../features/ColumnOrdering'
 import { ColumnPinning } from '../features/ColumnPinning'
 import { ColumnSizing } from '../features/ColumnSizing'
-import { ColumnVisibility } from '../features/ColumnVisibility'
+import { ColumnVisibility } from '../features/column-visibility/ColumnVisibility'
 import { GlobalFaceting } from '../features/GlobalFaceting'
 import { GlobalFiltering } from '../features/GlobalFiltering'
 import { RowExpanding } from '../features/RowExpanding'
