+++ packages/table-core/src/features/column-visibility/ColumnVisibility.types.ts
@@ -1,15 +1,4 @@
-import { ColumnPinningPosition } from '..'
-import {
-  Cell,
-  Column,
-  OnChangeFn,
-  Table,
-  Updater,
-  Row,
-  RowData,
-  TableFeature,
-} from '../types'
-import { getMemoOptions, makeStateUpdater, memo } from '../utils'
+import { Cell, Column, OnChangeFn, RowData, Updater } from '../../types'
 
 export type VisibilityState = Record<string, boolean>
 
@@ -145,173 +134,4 @@ export interface VisibilityColumn {
    * @link [Guide](https://tanstack.com/table/v8/docs/guide/column-visibility)
    */
   toggleVisibility: (value?: boolean) => void
-}
-
-//
-
-export const ColumnVisibility: TableFeature = {
-  _getInitialState: (state): VisibilityTableState => {
-    return {
-      columnVisibility: {},
-      ...state,
-    }
-  },
-
-  _getDefaultOptions: <TData extends RowData>(
-    table: Table<TData>
-  ): VisibilityDefaultOptions => {
-    return {
-      onColumnVisibilityChange: makeStateUpdater('columnVisibility', table),
-    }
-  },
-
-  _createColumn: <TData extends RowData, TValue>(
-    column: Column<TData, TValue>,
-    table: Table<TData>
-  ): void => {
-    column.toggleVisibility = value => {
-      if (column.getCanHide()) {
-        table.setColumnVisibility(old => ({
-          ...old,
-          [column.id]: value ?? !column.getIsVisible(),
-        }))
-      }
-    }
-    column.getIsVisible = () => {
-      const childColumns = column.columns
-      return (
-        (childColumns.length
-          ? childColumns.some(c => c.getIsVisible())
-          : table.getState().columnVisibility?.[column.id]) ?? true
-      )
-    }
-
-    column.getCanHide = () => {
-      return (
-        (column.columnDef.enableHiding ?? true) &&
-        (table.options.enableHiding ?? true)
-      )
-    }
-    column.getToggleVisibilityHandler = () => {
-      return (e: unknown) => {
-        column.toggleVisibility?.(
-          ((e as MouseEvent).target as HTMLInputElement).checked
-        )
-      }
-    }
-  },
-
-  _createRow: <TData extends RowData>(
-    row: Row<TData>,
-    table: Table<TData>
-  ): void => {
-    row._getAllVisibleCells = memo(
-      () => [row.getAllCells(), table.getState().columnVisibility],
-      cells => {
-        return cells.filter(cell => cell.column.getIsVisible())
-      },
-      getMemoOptions(table.options, 'debugRows', '_getAllVisibleCells')
-    )
-    row.getVisibleCells = memo(
-      () => [
-        row.getLeftVisibleCells(),
-        row.getCenterVisibleCells(),
-        row.getRightVisibleCells(),
-      ],
-      (left, center, right) => [...left, ...center, ...right],
-      getMemoOptions(table.options, 'debugRows', 'getVisibleCells')
-    )
-  },
-
-  _createTable: <TData extends RowData>(table: Table<TData>): void => {
-    const makeVisibleColumnsMethod = (
-      key: string,
-      getColumns: () => Column<TData, unknown>[]
-    ): (() => Column<TData, unknown>[]) => {
-      return memo(
-        () => [
-          getColumns(),
-          getColumns()
-            .filter(d => d.getIsVisible())
-            .map(d => d.id)
-            .join('_'),
-        ],
-        columns => {
-          return columns.filter(d => d.getIsVisible?.())
-        },
-        getMemoOptions(table.options, 'debugColumns', key)
-      )
-    }
-
-    table.getVisibleFlatColumns = makeVisibleColumnsMethod(
-      'getVisibleFlatColumns',
-      () => table.getAllFlatColumns()
-    )
-    table.getVisibleLeafColumns = makeVisibleColumnsMethod(
-      'getVisibleLeafColumns',
-      () => table.getAllLeafColumns()
-    )
-    table.getLeftVisibleLeafColumns = makeVisibleColumnsMethod(
-      'getLeftVisibleLeafColumns',
-      () => table.getLeftLeafColumns()
-    )
-    table.getRightVisibleLeafColumns = makeVisibleColumnsMethod(
-      'getRightVisibleLeafColumns',
-      () => table.getRightLeafColumns()
-    )
-    table.getCenterVisibleLeafColumns = makeVisibleColumnsMethod(
-      'getCenterVisibleLeafColumns',
-      () => table.getCenterLeafColumns()
-    )
-
-    table.setColumnVisibility = updater =>
-      table.options.onColumnVisibilityChange?.(updater)
-
-    table.resetColumnVisibility = defaultState => {
-      table.setColumnVisibility(
-        defaultState ? {} : table.initialState.columnVisibility ?? {}
-      )
-    }
-
-    table.toggleAllColumnsVisible = value => {
-      value = value ?? !table.getIsAllColumnsVisible()
-
-      table.setColumnVisibility(
-        table.getAllLeafColumns().reduce(
-          (obj, column) => ({
-            ...obj,
-            [column.id]: !value ? !column.getCanHide?.() : value,
-          }),
-          {}
-        )
-      )
-    }
-
-    table.getIsAllColumnsVisible = () =>
-      !table.getAllLeafColumns().some(column => !column.getIsVisible?.())
-
-    table.getIsSomeColumnsVisible = () =>
-      table.getAllLeafColumns().some(column => column.getIsVisible?.())
-
-    table.getToggleAllColumnsVisibilityHandler = () => {
-      return (e: unknown) => {
-        table.toggleAllColumnsVisible(
-          ((e as MouseEvent).target as HTMLInputElement)?.checked
-        )
-      }
-    }
-  },
-}
-
-export function _getVisibleLeafColumns<TData extends RowData>(
-  table: Table<TData>,
-  position?: ColumnPinningPosition | 'center'
-) {
-  return !position
-    ? table.getVisibleLeafColumns()
-    : position === 'center'
-      ? table.getCenterVisibleLeafColumns()
-      : position === 'left'
-        ? table.getLeftVisibleLeafColumns()
-        : table.getRightVisibleLeafColumns()
 }
